requirements.txt
notebook
numpy
pandas
scikit-learn
python-box
PyYAML
ensure
ucimlrepo
cassandra-driver
-e .